requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.116.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
oauthlib==3.3.1
openai==1.98.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.42
stack-data==0.6.3
starlette==0.47.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
WTForms==3.1.2